数据中台-移除日志写入SPI模块，精简代码

+++ auth-center/src/main/java/cn/cleanarch/dp/auth/endpoint/TokenEndpoint.java
@@ -186,7 +186,7 @@ public class TokenEndpoint {
 		int current = MapUtil.getInt(params, CommonConstants.CURRENT);
 		int size = MapUtil.getInt(params, CommonConstants.SIZE);
 		Set<String> keys = redisTemplate.keys(key);
-		List<String> pages = keys.stream().skip((current - 1) * size).limit(size).collect(Collectors.toList());
+		List<String> pages = keys.stream().skip((long) (current - 1) * size).limit(size).collect(Collectors.toList());
 		Page result = new Page(current, size);
 
 		List<TokenVO> tokenVoList = redisTemplate.opsForValue().multiGet(pages).stream().map(obj -> {
